requirements.txt
Flask==2.3.3
Pillow==10.2.0
Werkzeug==2.3.7
gunicorn==21.2.0
